commit
Theme :Land Usage Analysis in the South Zone of India

This project uses maps and data to study how land is being used in the South Zone. It helps us understand the balance between nature, farming, businesses, and homes. With this information, we can plan for the future in a way that benefits both people and the environment.